Move media assertion mixin to utils module
