+++ laces/tests/test_components.py
@@ -11,38 +11,13 @@ from django.test import SimpleTestCase
 from django.utils.safestring import SafeString
 
 from laces.components import Component, MediaContainer
+from laces.tests.utils import MediaAssertionMixin
 
 
 if TYPE_CHECKING:
     from typing import Any, Optional, Union
 
     from laces.typing import RenderContext
-
-
-class MediaAssertionMixin:
-    @staticmethod
-    def assertMediaEqual(first: widgets.Media, second: widgets.Media) -> bool:
-        """
-        Compare two `Media` instances.
-
-        The `Media` class does not implement `__eq__`, but its `__repr__` shows how to
-        recreate the instance.
-        We can use this to compare two `Media` instances.
-
-        Parameters
-        ----------
-        first : widgets.Media
-            First `Media` instance.
-        second : widgets.Media
-            Second `Media` instance.
-
-        Returns
-        -------
-        bool
-            Whether the two `Media` instances are equal.
-
-        """
-        return repr(first) == repr(second)
 
 
 class TestComponent(MediaAssertionMixin, SimpleTestCase):
